+++ packages/backend-canvas2d/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @glissade/backend-canvas2d
 
+## 0.35.0-pre.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @glissade/scene@0.35.0-pre.0
+  - @glissade/core@0.35.0-pre.0
+
 ## 0.34.0
 
 ### Patch Changes